+ some data moved inside plugin object (removed static vars from many .cpp) + correct working with memory (leaks removed) + containers refactoring
git-svn-id: http://filecopyex.googlecode.com/svn/trunk@42 aeb2dce1-986d-af1f-947f-f1de2426e1a6

+++ Framework/ObjString.h
@@ -70,7 +70,7 @@ public:
 	}
 
 	int AsInt() const { return _wtoi(ptr()); }
-	double AsFloat() const { return _wtof(ptr()); }
+	float AsFloat() const { return (float)_wtof(ptr()); }
 	bool AsBool() const { return (*this) == L"1"; } 
 	void ToUnicode(wchar_t* buf, size_t sz) const { CopyTo(buf, sz); }
 	void CopyTo(wchar_t* buf, size_t sz) const { wcscpy_s(buf, sz, ptr()); }
@@ -87,7 +87,7 @@ public:
 		_i64tow_s(v, buf, 64, 10);
 		str = buf;
 	}
-	explicit String(double v)
+	explicit String(float v)
 	{
 		wchar_t buf[64];
 		swprintf_s(buf, 64, L"%g", v);
